threads: add heap type tests
This change adds generated tests for all of the abstract heap types when
they are `shared` and placed in `shared` globals. Since globals are the
only thing we can mark shared now that can touch these types, they are a
good vehicle for testing the parsing, encoding, and validation of all of
this. Eventually the same must be done for concrete heap types once
composite types (`func`, `struct`, `array`) can be marked shared.